+++ android/src/main/java/com/dexterous/flutterlocalnotifications/NotificationStyle.java
@@ -2,6 +2,7 @@ package com.dexterous.flutterlocalnotifications;
 
 public enum NotificationStyle{
     Default,
+    BigPicture,
     BigText,
     Inbox
 }